requirements.txt
langchain 
chromadb 
sentence-transformers 
streamlit 
pypdf 
pydantic 
torch 
llama-cpp-python